+++ src/RichDomainStore.Catalog.Application/Dtos/UpdateProductDto.cs
@@ -16,10 +16,10 @@ namespace RichDomainStore.Catalog.Application.Dtos
 
         public string Image { get; set; }
 
-        public int Height { get; set; }
+        public decimal Height { get; set; }
 
-        public int Width { get; set; }
+        public decimal Width { get; set; }
 
-        public int Depth { get; set; }
+        public decimal Depth { get; set; }
     }
 }
